# src/main/java/ru/skypro/homework/dto/NewPasswordDto.java
package ru.skypro.homework.dto;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class NewPasswordDto {

    private String currentPassword;

    private String newPasswordDto;

}